Generalise `re_replacement_seq` to deal with special symbols
This PR is similar to https://github.com/dottxt-ai/outlines-core/pull/90, and 
generalises the regex to deal with all the previous, and hopefully all future cases as well.

The new special case not covered by the previous approach are the `�?`
and `�,` tokens, used by Salamandra models. Since all these special
tokens (new and old) consist of one or more � symbols, with an optional
single-character prefix and/or suffix, we can simplify and generalise
the pattern to r"^.?�+.?$".