Use pypa to publish to pypi instead of maturin (#207)
As we are having issues with the `maturin upload` action, this PR
proposes to use the pypa action that is used in Outlines to publish the
wheels to pypi. It also adds a debug step that prints out some
information on the working directory and that are intended to be removed
in the future.