Allow building the lib for more flavors (#200)
The library has limited support for embedded environments, such as iOS
or Mac Catalyst, all due to the huge dependency graph that hf-hub has.

This change makes it possible to strip the hf-hub dependency. The
current use for the hf-hub is preserved as a default feature.

Tested to work with:
* `cargo build --release --target aarch64-apple-ios
--no-default-features`
* `cargo build --release --target aarch64-apple-ios-macabi
--no-default-features`